Move exceptions and constants first take
