+++ sparkmagic/sparkmagic/magics/__init__.py
@@ -1,2 +1,1 @@
-__version__ = '0.0.1'
 from sparkmagic.magics.remotesparkmagics import *
