+++ sparkmagic/sparkmagic/utils/constants.py
